authentication/firebase/python/src/requirements.txt
firebase_admin==5.2.0
requests==2.27.1
urllib3==1.25.9
Flask==1.0.3
python-dotenv==0.20.0